Move logUserIn and submitNewUser methods out of App; create NewUserForm component; add logUserIn method to LoginForm and submitNewUser method to NewUserForm
