+++ src/components/LoginForm/LoginForm.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import { createNewUser, loginUser } from '../../apiCalls'
-import { Link, Redirect } from 'react-router-dom';
+import { loginUser } from '../../apiCalls'
 
 class LoginForm extends Component {
     constructor() {
@@ -13,13 +12,13 @@ class LoginForm extends Component {
             loginPassWord: ''
         }
     }
-
-    submitNewUserInfo = event => {
-        event.preventDefault()
-        const { newName, newEmail, newPassword } = this.state
-        const newUser = {name: newName, email: newEmail, password: newPassword}
-        this.props.addNewUser(newUser);
-        this.setState({newName: '', newEmail: '', newPassword: ''})
+    
+    logUserIn = async (userInfo) => {
+        try {
+          await loginUser(userInfo);
+        } catch({ message }) {
+          this.setState({error: message})
+        }
     }
 
     handleChange = (e) => {
@@ -30,25 +29,13 @@ class LoginForm extends Component {
         e.preventDefault();
         const { loginEmail, loginPassword } = this.state;
         const user = { email: loginEmail, password: loginPassword};
-        this.props.logUserIn(user);
-
-
+        this.logUserIn(user);
         this.setState({ loginEmail: '', loginPassword: ''})
     }
     
     render() {
         return(
             <div>
-                <section className="create-account">
-                    <h3>CREATE ACCOUNT</h3>
-                    <form>
-                        <input type="text" placeholder="insert your name" name="newName" value={this.state.newName} onChange={this.handleChange} />
-                        <input type="email" placeholder="insert your e-mail" name="newEmail" value={this.state.newEmail} onChange={this.handleChange} />
-                        <input type="password" placeholder="insert your password" name="newPassword" value={this.state.newPassword} onChange={this.handleChange} />
-                        <button className="create-user-button" onClick={(e) => this.submitNewUserInfo(e)}>SUBMIT</button>
-                    </form>
-                </section>
-                <h3>OR</h3>
                 <section className="login">
                     <h3>LOG IN</h3>
                     <form>
