+++ src/containers/App/App.js
@@ -1,42 +1,21 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
 import './App.scss';
 import MoviesContainer from '../MoviesContainer/MoviesContainer';
 import NavBar from '../../components/NavBar/NavBar';
-import { saveMovies } from '../../actions';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import { fetchMovies } from '../../apiCalls.js';
 import LoginForm from '../../components/LoginForm/LoginForm';
-import { createNewUser, loginUser } from '../../apiCalls'
+import { saveMovies } from '../../actions';
+import { fetchMovies } from '../../apiCalls.js';
 
 class App extends Component {
   async componentDidMount() {
-    const { movies } = this.props; 
-    console.log(movies)
     try {
       const movieData = await fetchMovies();
       await this.props.saveMovies(movieData)
     }
-
-    catch {
-    }
-  }
-
-  addNewUser = async (newUserInfo) => {
-    try {
-      await createNewUser(newUserInfo);
-    } catch({ message }) {
-      this.setState({error: message})
-    }
-  }
-
-  logUserIn = async (userInfo) => {
-    try {
-      await loginUser(userInfo);
-    } catch({ message }) {
-      this.setState({error: message})
-    }
+    catch {   }
   }
 
   render() {
@@ -44,7 +23,7 @@ class App extends Component {
       <div className="App">
         <NavBar />
         <Route exact path='/' render={() => <MoviesContainer /> } />
-        <Route exact path='/login' render={() => <LoginForm addNewUser={this.addNewUser} logUserIn={this.logUserIn} />} />
+        <Route exact path='/login' render={() => <LoginForm />} />
       </div>
     )
   }
